Reorganize binding code
Reorganize the bindings into header and implementation
