+++ vital/bindings/python/vital/algo/algorithm.cxx
@@ -27,56 +27,19 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-
-/**
- * \file algorithm_implementation.cxx
- *
- * \brief python bindings for algorithm
- */
-
-
 #include <pybind11/pybind11.h>
-#include <pybind11/stl.h>
-
 #include <vital/algo/algorithm.h>
-#include <vital/algo/image_object_detector.h>
-#include <vital/algo/trampoline/image_object_detector_trampoline.txx>
-#include <sstream>
+#include <vital/bindings/python/vital/algo/trampoline/algorithm_trampoline.txx>
+#include <vital/bindings/python/vital/algo/algorithm.h>
 
 namespace py = pybind11;
 
-void algorithm_implementation(py::module &m);
-
-template<class implementation, class trampoline>
-void register_algorithm_implementation(py::module &m,
-                                      const std::string implementation_name)
+void algorithm(py::module &m)
 {
-  std::stringstream impl_name;
-  impl_name << "_algorithm<" << implementation_name << ">";
-
-  py::class_< kwiver::vital::algorithm_def<implementation>,
-              std::shared_ptr<kwiver::vital::algorithm_def<implementation>>,
-              kwiver::vital::algorithm,
-              trampoline >(m, impl_name.str().c_str())
-    .def(py::init())
-    .def_static("create", &kwiver::vital::algorithm_def<implementation>::create)
-    .def_static("registered_names",
-                &kwiver::vital::algorithm_def<implementation>::registered_names)
-    .def_static("get_nested_algo_configuration",
-                &kwiver::vital::algorithm_def<implementation>::get_nested_algo_configuration)
-    .def_static("set_nested_algo_configuration",
-                &kwiver::vital::algorithm_def<implementation>::set_nested_algo_configuration)
-    .def_static("check_nested_algo_configuration",
-                &kwiver::vital::algorithm_def<implementation>::check_nested_algo_configuration);
+  py::class_<kwiver::vital::algorithm, std::shared_ptr<kwiver::vital::algorithm>,
+             algorithm_trampoline<>>(m, "_algorithm")
+    .def("get_configuration", &kwiver::vital::algorithm::get_configuration)
+    .def("set_configuration", &kwiver::vital::algorithm::set_configuration)
+    .def("check_configuration", &kwiver::vital::algorithm::check_configuration);
 }
 
-void image_object_detector_implementation(py::module &m);
-
-
-PYBIND11_MODULE(algorithm, m)
-{
-  algorithm_implementation(m);
-  register_algorithm_implementation<kwiver::vital::algo::image_object_detector, 
-                                    py_iod_algorithm_def<>>(m, "image_object_detector");
-  image_object_detector_implementation(m);
-}
