+++ vital/bindings/python/vital/algo/image_object_detector.h
@@ -1,0 +1,9 @@
+#ifndef KWIVER_VITAL_PYTHON_IMAGE_OBJECT_DETECTOR_H_
+#define KWIVER_VITAL_PYTHON_IMAGE_OBJECT_DETECTOR_H_
+
+#include <pybind11/pybind11.h>
+
+namespace py = pybind11;
+
+void image_object_detector(py::module &m);
+#endif
